实现回复列表功能

+++ src/main/java/com/firstweb/demo/pojo/QuestionPOJO.java
@@ -9,13 +9,13 @@ import lombok.Data;
  */
 @Data
 public class QuestionPOJO {
-    private Integer id;
+    private long id;
     private String title;
     private String description;
     private String tag;
     private long gmtCreate;
     private long gmtModified;
-    private Integer creator;
+    private long creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
